shopping list react component
