+++ App.tsx
@@ -1,33 +1,13 @@
 import { StatusBar } from "expo-status-bar";
-import { Alert, StyleSheet, Text, TouchableOpacity, View } from "react-native";
-import { theme } from "./theme";
+import { StyleSheet, View } from "react-native";
+import { ShoppingListItem } from "./components/ShoppingList";
 
 export default function App() {
-  const handleDelete = () => {
-    Alert.alert(
-      "Are you sure you want to delete this?",
-      "It will be gone for good",
-      [
-        {
-          text: "Yes",
-          onPress: () => console.log("Ok, deleting"),
-          style: "destructive",
-        },
-        {
-          text: "Cancel",
-          style: "cancel",
-        }
-      ])
-  }
-
   return (
     <View style={styles.container}>
-      <View style={styles.itemsContainer}>
-        <Text style={styles.itemText}>Coffee</Text>
-        <TouchableOpacity style={styles.button} onPress={handleDelete} activeOpacity={0.8}>
-          <Text style={styles.buttonText}>Delete</Text>
-        </TouchableOpacity>
-      </View>
+      <ShoppingListItem name="Coffee"></ShoppingListItem>
+      <ShoppingListItem name="Tea"></ShoppingListItem>
+      <ShoppingListItem name="Chocolate"></ShoppingListItem>
     </View>
   );
 }
@@ -38,24 +18,4 @@ const styles = StyleSheet.create({
     backgroundColor: "#fff",
     justifyContent: "center",
   },
-  itemsContainer: {
-    borderBottomWidth: 1,
-    borderBottomColor: "#1a759f",
-    paddingHorizontal: 8,
-    paddingVertical: 16,
-    flexDirection: "row",
-    alignItems: "center",
-    justifyContent: "space-between",
-  },
-  itemText: { fontSize: 18, fontWeight: "200" },
-  button: {
-    backgroundColor: theme.colorBlack,
-    padding: 8,
-    borderRadius: 6,
-  },
-  buttonText: {
-    color: theme.colorWhite,
-    fontWeight: "bold",
-    textTransform: "uppercase"
-  },
-}); 
+});
