+++ package.json
@@ -22,6 +22,7 @@
     "eslint-config-expo": "~8.0.1",
     "eslint-config-prettier": "^9.1.0",
     "eslint-plugin-prettier": "^5.2.1",
+    "eslint-plugin-react-native": "^5.0.0",
     "prettier": "^3.4.2",
     "typescript": "^5.3.3"
   },
