+++ theme.ts
@@ -1,7 +1,7 @@
-// We can use https://www.nativewind.dev/ which is Tailwind for React-Native 
+// We can use https://www.nativewind.dev/ which is Tailwind for React-Native
 
 export const theme = {
-    colorCerulean: "#1a759f",
-    colorWhite: "#fff",
-    colorBlack: "#000",
-}
+  colorCerulean: "#1a759f",
+  colorWhite: "#fff",
+  colorBlack: "#000",
+};
